Moved methods to respective classes for file and fixed specs
