+++ lib/train/file/remote/aix.rb
@@ -1,0 +1,21 @@
+# encoding: utf-8
+
+require 'train/file/remote/unix'
+
+module Train
+  class File
+    class Remote
+      class Aix < Train::File::Remote::Unix
+        def link_path
+          return nil unless symlink?
+          @link_path ||=
+            @backend.run_command("perl -e 'print readlink shift' #{@spath}").stdout.chomp
+        end
+
+        def mounted
+          @mounted ||= @backend.run_command("lsfs -c #{@spath}")
+        end
+      end
+    end
+  end
+end
