+++ lib/train/transports/winrm_connection.rb
@@ -51,7 +51,7 @@ class Train::Transports::WinRM
     end
 
     def file(path)
-      @files[path] ||= WindowsFile.new(self, path)
+      @files[path] ||= Train::File::Remote::Windows.new(self, path)
     end
 
     def run_command(command)
