+++ test/integration/tests/path_missing_test.rb
@@ -69,9 +69,8 @@ describe 'file interface' do
       file.product_version.must_be_nil
     end
 
-    it 'has no file_version' do
+    it 'has no product_version' do
       file.file_version.must_be_nil
     end
-
   end
 end
