+++ test/integration/tests/path_symlink_test.rb
@@ -85,11 +85,11 @@ describe 'file interface' do
     end
 
     it 'has no product_version' do
-      file.product_version.must_equal(nil)
+      file.product_version.must_be_nil
     end
 
-    it 'has no file_version' do
-      file.file_version.must_equal(nil)
+    it 'has no product_version' do
+      file.file_version.must_be_nil
     end
   end
 end
