+++ test/unit/extras/os_detect_windows_test.rb
@@ -97,7 +97,7 @@ describe 'os_detect_windows' do
       detector.detect_windows
       detector.platform[:family].must_equal('windows')
       detector.platform[:name].must_equal('Windows 4.10.1998')
-      detector.platform[:arch].must_equal(nil)
+      detector.platform[:arch].must_be_nil
       detector.platform[:release].must_equal('4.10.1998')
     end
   end
