+++ test/windows/winrm_test.rb
@@ -45,6 +45,131 @@ describe 'windows winrm command' do
     cmd.stderr.must_equal ''
   end
 
+
+  describe 'file' do
+    before do
+      @temp = Tempfile.new('foo')
+      @temp.write("hello world")
+      @temp.rewind
+    end
+
+    let(:file) { conn.file(@temp.path) }
+
+    it 'exists' do
+      file.exist?.must_equal(true)
+    end
+
+    it 'is a file' do
+      file.file?.must_equal(true)
+    end
+
+    it 'has type :file' do
+      file.type.must_equal(:file)
+    end
+
+    it 'has content' do
+      # TODO: this shouldn't include newlines that aren't in the original file
+      file.content.must_equal("hello world\r\n")
+    end
+
+    it 'has owner name' do
+      file.owner.wont_be_nil
+    end
+
+    it 'has no group name' do
+      file.group.must_be_nil
+    end
+
+    it 'has no mode' do
+      file.mode.must_be_nil
+    end
+
+    it 'has an md5sum' do
+      file.md5sum.wont_be_nil
+    end
+
+    it 'has an sha256sum' do
+      file.sha256sum.wont_be_nil
+    end
+
+    it 'has no modified time' do
+      file.mtime.must_be_nil
+    end
+
+    it 'has no size' do
+      file.size.wont_be_nil
+    end
+
+    it 'has size 11' do
+      size = ::File.size(@temp)
+      file.size.must_equal size
+    end
+
+    it 'has no selinux_label handling' do
+      file.selinux_label.must_be_nil
+    end
+
+    it 'has product_version' do
+      file.product_version.wont_be_nil
+    end
+
+    # TODO: This is not failing in manual testing
+    # it 'returns basname of file' do
+    #   basename = ::File.basename(@temp)
+    #   file.basename.must_equal basename
+    # end
+
+    it 'has file_version' do
+      file.file_version.wont_be_nil
+    end
+
+    it 'returns nil for mounted' do
+      file.mounted.must_be_nil
+    end
+
+    it 'has no link_path' do
+      file.link_path.must_be_nil
+    end
+
+    it 'has no uid' do
+      file.uid.must_be_nil
+    end
+
+    it 'has no gid' do
+      file.gid.must_be_nil
+    end
+
+    it 'provides a json representation' do
+      j = file.to_json
+      j.must_be_kind_of Hash
+      j['type'].must_equal :file
+    end
+
+    after do
+      @temp.close
+      @temp.unlink
+    end
+  end
+
+  describe 'file' do
+    before do
+      @temp = Tempfile.new('foo bar')
+      @temp.write("hello world")
+      @temp.rewind
+    end
+
+    let(:file) { conn.file(@temp.path) }
+
+    it 'provides the full path with whitespace for path #{@temp.path}' do
+      file.path.must_equal @temp.path
+    end
+
+    after do
+      @temp.close
+      @temp.unlink
+    end
+  end
+
   after do
     # close the connection
     conn.close
